requirements.txt
streamlit
tensorflow
numpy
Pillow
scikit-learn